requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via slackclient
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
attrs==23.2.0
    # via
    #   aiohttp
    #   outcome
    #   trio
babel==2.15.0
    # via -r requirements.in
bcrypt==4.1.3
    # via paramiko
beautifulsoup4==4.12.3
    # via bs4
billiard==4.2.0
    # via celery
blinker==1.8.2
    # via flask
bs4==0.0.2
    # via -r requirements.in
cachelib==0.9.0
    # via flask-caching
cachetools==5.3.3
    # via google-auth
celery==5.4.0
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   httpx
    #   requests
    #   selenium
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
    #   trio
chardet==3.0.4
    # via httpx
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
cryptography==42.0.7
    # via paramiko
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-caching
    #   flask-sqlalchemy
flask-caching==2.3.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.3
    # via -r requirements.in
fuzzywuzzy==0.18.0
    # via -r requirements.in
google-api-core==2.19.0
    # via google-api-python-client
google-api-python-client==2.129.0
    # via -r requirements.in
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
googleapis-common-protos==1.63.0
    # via google-api-core
googletrans==3.0.0
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
h11==0.9.0
    # via
    #   httpcore
    #   wsproto
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2024.5.1
    # via httpx
httpcore==0.9.1
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
httpx==0.13.3
    # via googletrans
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   requests
    #   trio
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
kombu==5.3.7
    # via celery
levenshtein==0.25.1
    # via python-levenshtein
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
orderedmultidict==1.0.1
    # via furl
outcome==1.3.0.post0
    # via trio
paramiko==3.4.0
    # via sshtunnel
phonenumbers==8.13.37
    # via -r requirements.in
prompt-toolkit==3.0.43
    # via click-repl
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pycountry==23.12.11
    # via -r requirements.in
pycparser==2.22
    # via cffi
pymysql==1.1.0
    # via -r requirements.in
pynacl==1.5.0
    # via paramiko
pyparsing==3.1.2
    # via httplib2
pysocks==1.7.1
    # via urllib3
python-anticaptcha==1.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via celery
python-levenshtein==0.25.1
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
rapidfuzz==3.9.1
    # via levenshtein
redis==5.0.4
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   python-anticaptcha
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rfc3986==1.5.0
    # via httpx
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
selenium==4.21.0
    # via -r requirements.in
six==1.16.0
    # via
    #   furl
    #   oauth2client
    #   orderedmultidict
    #   python-anticaptcha
    #   python-dateutil
slackclient==2.9.4
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   httpcore
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.30
    # via flask-sqlalchemy
sshtunnel==0.4.0
    # via -r requirements.in
tld==0.13
    # via -r requirements.in
trio==0.25.1
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.11.0
    # via
    #   selenium
    #   sqlalchemy
tzdata==2024.1
    # via celery
unidecode==1.3.8
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3[socks]==2.2.1
    # via
    #   requests
    #   selenium
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==3.0.3
    # via flask
wsproto==1.2.0
    # via trio-websocket
yarl==1.9.4
    # via aiohttp
